Rename interface files
